Reduce calls to malloc in unordered_map (#96)
Reduce the number of malloc calls in unordered_map. Improves
efficiency by 50% (meaning it takes half the time that it used to).

+++ src/include/unordered_map.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2019 Bailey Thompson
+ * Copyright (c) 2017-2020 Bailey Thompson
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -40,7 +40,7 @@ unordered_map unordered_map_init(size_t key_size,
 
 /* Utility */
 int unordered_map_rehash(unordered_map me);
-int unordered_map_size(unordered_map me);
+size_t unordered_map_size(unordered_map me);
 int unordered_map_is_empty(unordered_map me);
 
 /* Accessing */
